Fix hard break for reactive messaging fat
