+++ dev/com.ibm.ws.microprofile.reactive.messaging.kafka/src/com/ibm/ws/microprofile/reactive/messaging/kafka/KafkaIncomingConnector.java
@@ -1,14 +1,11 @@
 /*******************************************************************************
- * Copyright (c) 2019, 2023 IBM Corporation and others.
+ * Copyright (c) 2019, 2026 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License 2.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-2.0/
  *
  * SPDX-License-Identifier: EPL-2.0
- *
- * Contributors:
- *     IBM Corporation - initial API and implementation
  *******************************************************************************/
 package com.ibm.ws.microprofile.reactive.messaging.kafka;
 
@@ -161,8 +158,8 @@ public class KafkaIncomingConnector implements IncomingConnectorFactory, Quiesce
             }
 
             // Create our connector around the kafkaConsumer
-            KafkaInput<String, Object> kafkaInput = new KafkaInput<>(this.kafkaAdapterFactory, partitionTrackerFactory, kafkaConsumer, asyncProvider,
-                                                                     topic, unackedLimit, fastAck);
+            KafkaInput<String, Object> kafkaInput = new AppAwareKafkaInput<>(this.kafkaAdapterFactory, partitionTrackerFactory, kafkaConsumer, asyncProvider,
+                                                                             topic, unackedLimit, fastAck);
             kafkaInputs.add(kafkaInput);
 
             return kafkaInput.getPublisher();
